+++ HISTORY.md
@@ -18,6 +18,8 @@
 - Sort MACS2 bedgraph files before compressing. Thanks to @LMannarino for the suggestion.
 - Check for the reserved field `sample` in RNA-seq metadata and quit with a useful error message. 
   Thanks to @marypiper for suggesting this.
+- Split ATAC-seq BAM files into nucleosome-free and mono/di/tri nucleosome files, so we can call 
+  peaks on them separately.
 
 ## 1.1.8 (28 October 2019)
 - Add `antibody` configuration option. Setting a specific antibody for ChIP-seq will use appropriate